third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/watch-expressions-preserve-expansion.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

var globalObject = {
    foo: {
        bar: {
            baz: 2012
        }
    }
};
var windowAlias = window;

var test = function()
{
    var watchExpressionsPane = WebInspector.panels.sources.sidebarPanes.watchExpressions;
    watchExpressionsPane.expand();
    watchExpressionsPane.addExpression("globalObject");
    watchExpressionsPane.addExpression("windowAlias");
    InspectorTest.runAfterPendingDispatches(step2);

    function step2()
    {
        InspectorTest.addResult("Watch expressions added.");
        expandWatchExpression(["globalObject", "foo", "bar"], step3);
    }

    function step3()
    {
        InspectorTest.addResult("Watch expressions expanded.");
        dumpWatchExpressions();
        InspectorTest.reloadPage(step4);
    }

    function step4()
    {
        InspectorTest.addResult("Watch expressions after page reload:");
        dumpWatchExpressions();
        InspectorTest.completeTest();
    }

    function dumpWatchExpressions()
    {
        var pane = WebInspector.panels.sources.sidebarPanes.watchExpressions;

        for (var i = 0; i < pane._watchExpressions.length; i++) {
            var watch = pane._watchExpressions[i];
            InspectorTest.addResult(watch.expression() + ": " + watch._objectPropertiesSection._object._description);
            dumpObjectPropertiesTreeElement(watch._objectPropertiesSection.objectTreeElement(), "  ");
        }
    }

    function dumpObjectPropertiesTreeElement(treeElement, indent)
    {
        if (treeElement.property)
            InspectorTest.addResult(indent + treeElement.property.name + ": " + treeElement.property.value._description);
        for (var i = 0; i < treeElement.children().length; i++)
            dumpObjectPropertiesTreeElement(treeElement.children()[i], "  " + indent);
    }

    function expandProperties(treeoutline, path, callback)
    {
        treeoutline.addEventListener(TreeOutline.Events.ElementAttached, elementAttached);
        treeoutline.expand();

        function elementAttached(event)
        {
            var treeElement = event.data;
            if (treeElement.property.name !== path[0])
                return;

            var childName = path.shift();
            InspectorTest.addResult("expanded " + childName + " " + treeElement.property.value);

            if (path.length) {
                treeElement.expand();
                return;
            }

            treeoutline.removeEventListener(TreeOutline.Events.ElementAttached, elementAttached);
            callback();
        }
    }

    function expandWatchExpression(path, callback)
    {
        var pane = WebInspector.panels.sources.sidebarPanes.watchExpressions;
        var expression = path.shift();
        for (var i = 0; i < pane._watchExpressions.length; i++) {
            var watch = pane._watchExpressions[i];
            if (watch.expression() === expression) {
                expandProperties(watch._objectPropertiesSection, path, callback);
                break;
            }
        }

    }
}

</script>
</head>
<body onload="runTest()">
<p>Test that watch expressions expansion state is restored after update.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=99304">Bug 99304</a>
</body>
</html>
