third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/watch-expressions-preserve-expansion-expected.txt

Test that watch expressions expansion state is restored after update.

Bug 99304
Watch expressions added.
expanded foo [object Object]
expanded bar [object Object]
Watch expressions expanded.
globalObject: Object
    foo: Object
      bar: Object
    __proto__: Object
windowAlias: Window
Page reloaded.
Watch expressions after page reload:
globalObject: Object
    foo: Object
      bar: Object
      __proto__: Object
    __proto__: Object
windowAlias: Window
